refactor: Introduce DTOs to prevent entity exposure
Replace direct usage of Ticket entity in controller and service layers
with dedicated DTOs (TicketCreateRequestDTO, TicketResponseDTO,
TicketUpdateRequestDTO). This prevents exposing internal entity
structures to API consumers and improves separation of concerns.

Additionally:
- Change ticket price field from Double to BigDecimal for precision.
- Update .gitignore to exclude repomix output files.
- Remove test endpoint for expired reservation in controller.

Fixes #2

+++ src/main/java/com/omniticket/reservation_service/service/EmailService.java
@@ -1,5 +1,7 @@
 package com.omniticket.reservation_service.service;
 
+import java.math.BigDecimal;
+
 import jakarta.mail.MessagingException;
 import jakarta.mail.internet.MimeMessage;
 import lombok.RequiredArgsConstructor;
@@ -15,7 +17,7 @@ public class EmailService {
 
     private final JavaMailSender mailSender;
 
-    public void sendTicketEmail(String to, String seat, Double price) throws MessagingException {
+    public void sendTicketEmail(String to, String seat, BigDecimal price) throws MessagingException {
         MimeMessage mimeMessage = mailSender.createMimeMessage();
         MimeMessageHelper helper = new MimeMessageHelper(mimeMessage, "utf-8");
